+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
 ### FEATURES
 
+- `[config]` Add EventBusBufferCapacity setting.
+  ([\#5849](https://github.com/cometbft/cometbft/pull/5849))
 ### STATE-BREAKING
 
 ### API-BREAKING